+++ test/integration/distributed/setup/node_test.go
@@ -55,7 +55,8 @@ var _ = Describe("Node registration", func() {
 			"--etcd-endpoints", etcdEndpoint,
 			"--node-host-provider", "flag",
 			"--node-host", nodeHost,
-			"--schema-registry-mode=etcd")
+			"--schema-registry-mode=etcd",
+			"--node-discovery-mode=etcd")
 		Eventually(helpers.HTTPHealthCheck(httpAddr, ""), flags.EventuallyTimeout).Should(Succeed())
 		Eventually(func() (map[string]*databasev1.Node, error) {
 			return helpers.ListKeys(etcdEndpoint, fmt.Sprintf("/%s/nodes/%s:%d", namespace, nodeHost, ports[2]))
@@ -78,7 +79,8 @@ var _ = Describe("Node registration", func() {
 			"--etcd-endpoints", etcdEndpoint,
 			"--node-host-provider", "flag",
 			"--node-host", nodeHost,
-			"--schema-registry-mode=etcd")
+			"--schema-registry-mode=etcd",
+			"--node-discovery-mode=etcd")
 		Eventually(
 			helpers.HealthCheck(addr, 10*time.Second, 10*time.Second, grpc.WithTransportCredentials(insecure.NewCredentials())),
 			flags.EventuallyTimeout).Should(Succeed())